requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.15
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiosqlite==0.19.0
gitpython==3.1.31
pytest==7.3.1
httpx==0.24.1
pytest-asyncio==0.21.1
pytest-cov==3.0.0